backend/requirements.txt
tqdm
torch
opencv-python
ffmpegcv
numpy
python_speech-features
scipy
scenedetect
scikit-learn
gdown # model weights
torchvision
pandas
transformers
accelerate # multi device inference
datasets
google-genai 
pysubs2 # subtitle files
boto3
fastapi[standard] # expose endpoints
whisperx # speech detect
